pyproject.toml
[tool.poetry]
name = "apptwitter"
version = "0.1.0"
description = "Aplicación local para automatizar difusión en X (Twitter) de artículos y generar tweets de engagement"
authors = ["Federico"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
click = "^8.1.7"
python-dotenv = "^1.0.0"
pyyaml = "^6.0.1"
tweepy = "^4.14.0"
rapidfuzz = "^3.6.1"
rich = "^13.7.0"
python-dateutil = "^2.8.2"
openai = {version = "^1.10.0", optional = true}
anthropic = {version = "^0.18.0", optional = true}
google-generativeai = {version = "^0.3.0", optional = true}

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
black = "^24.1.1"
ruff = "^0.1.14"

[tool.poetry.extras]
llm-openai = ["openai"]
llm-anthropic = ["anthropic"]
llm-gemini = ["google-generativeai"]

[tool.poetry.scripts]
app = "src.cli:cli"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
